# openapi-order-module.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "营养餐订单系统 API",
    "description": "营养餐订单管理系统的API接口规范，包含地址管理、订单管理和配送管理",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
...
  "paths": {
    "/addresses": {
      "get": {
        "tags": ["地址管理"],
        "summary": "获取用户的收货地址列表",
        "description": "返回当前登录用户的所有收货地址",
        "operationId": "listAddresses",
        "responses": {
          "200": {
            "description": "成功获取地址列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["地址管理"],
        "summary": "添加收货地址",
        "description": "添加新的收货地址",
        "operationId": "addAddress",
        "requestBody": {
          "description": "地址信息",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "地址添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/addresses/{id}": {
      "get": {
        "tags": ["地址管理"],
        "summary": "获取地址详情",
        "description": "根据地址ID获取地址详情",
        "operationId": "getAddressDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "地址ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取地址详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "地址不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["地址管理"],
        "summary": "更新收货地址",
        "description": "根据地址ID更新收货地址",
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "地址ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "地址信息",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "地址更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "地址不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["地址管理"],
        "summary": "删除收货地址",
        "description": "根据地址ID删除收货地址",
        "operationId": "deleteAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "地址ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "地址删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "地址不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/addresses/{id}/default": {
      "put": {
        "tags": ["地址管理"],
        "summary": "设置默认收货地址",
        "description": "将指定地址设置为默认收货地址",
        "operationId": "setDefaultAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "地址ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置默认地址成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "地址不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "地址ID"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "receiver": {
            "type": "string",
            "description": "收货人姓名"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "province": {
            "type": "string",
            "description": "省份"
          },
          "city": {
            "type": "string",
            "description": "城市"
          },
          "district": {
            "type": "string",
            "description": "区/县"
          },
          "detailAddress": {
            "type": "string",
            "description": "详细地址"
          },
          "lat": {
            "type": "number",
            "format": "double",
            "description": "纬度"
          },
          "lng": {
            "type": "number",
            "format": "double",
            "description": "经度"
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否默认地址"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "AddressRequest": {
        "type": "object",
        "required": [
          "receiver",
          "phone",
          "province",
          "city",
          "district",
          "detailAddress"
        ],
        "properties": {
          "receiver": {
            "type": "string",
            "description": "收货人姓名"
          },
          "phone": {
            "type": "string",
            "description": "联系电话",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "province": {
            "type": "string",
            "description": "省份"
          },
          "city": {
            "type": "string",
            "description": "城市"
          },
          "district": {
            "type": "string",
            "description": "区/县"
          },
          "detailAddress": {
            "type": "string",
            "description": "详细地址"
          },
          "lat": {
            "type": "number",
            "format": "double",
            "description": "纬度"
          },
          "lng": {
            "type": "number",
            "format": "double",
            "description": "经度"
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否默认地址",
            "default": false
          }
        }
      },
      "AddressListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          }
        }
      },
      "AddressResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/Address"
          }
        }
      },
      "AddressIdResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "integer",
            "format": "int64",
            "description": "地址ID"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "null",
            "description": "无数据"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功",
            "default": false
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "type": "null",
            "description": "无数据"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 